\name{flowUtils-package}
\alias{flowUtils-package}
\alias{flowUtils}
\docType{package}
\title{
  Utilities for flow cytometry data
}
\description{
  Utilities to process flow cytometry data. This package
  heavily depends on the flowCore package.
}
\details{
\tabular{ll}{
Package: \tab flowUtils\cr
Type: \tab Package\cr
Version: \tab 0.2.1\cr
Date: \tab 2006-11-16\cr
License: \tab Artistic\cr
}
The main features of this package provide compatability to the data
standards in the gatingML, transformationML and compensationML markup
languages.
}
\author{


  Maintainer: Florian Hahne <f.hahne@dkfz.de>
  Authors: M. Dalphin, B. Ellis, R. Gentleman, F. Hahne, N. Le Meur, B. Purcell
}


\keyword{ package }
\seealso{
\code{\link[flowCore:flowCore-package]{flowCore}}
}
\examples{
## examples go here
}
