\name{testGatingCompliance}
\alias{testGatingCompliance}

\title{
       Function to perfom all the gating-ML Compliance tests
      }
\description{
  This funtion performs all the gating-ML compliance tests .The results obtained
  are compared with the expected results and an html report is generated.
  The data files and XML files required for the tests are provided in the 
  gatingML data package.
 }
\usage{
 testGatingCompliance(file)
}

\arguments{
  \item{file}{name of the file in which the report generated is to be saved }
  }
}
\details{

}

\references{ ~put references to the literature/web site here ~ }
\author{Gopalakrishnan N}
\seealso{\code{\link[flowCore]{filterSet-class}}}
\examples{


}

\keyword{}

